fix test bugs
